Add default sizes to the nodes in the example to avoid confusion when used with ZoomToFit
As the nodes did not have sizes, ZoomToFit behaved erroneously as the GetBounds method returned ♾ for its width & height elements. This is subtle, and super-confusing, so to avoid someone else having this future pain, I've updated the Layout demo code to provide some default sizes that match the default expectations in the layout code. 

This means the demo can be simply extended to include a ZoomToFit call with no ill effects